[snomed] use short for referencedComponentType in SnomedConceptDocument

+++ snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/datastore/index/entry/SnomedConceptDocument.java
@@ -234,7 +234,7 @@ public class SnomedConceptDocument extends SnomedComponentDocument implements IT
 		private LongSet statedParents;
 		private LongSet statedAncestors;
 		private SnomedRefSetType refSetType;
-		private int referencedComponentType;
+		private short referencedComponentType;
 		private int mapTargetComponentType;
 		private float doi = DEFAULT_DOI;
 		private long refSetStorageKey = CDOUtils.NO_STORAGE_KEY;
@@ -328,7 +328,7 @@ public class SnomedConceptDocument extends SnomedComponentDocument implements IT
 			return getSelf();
 		}
 
-		Builder referencedComponentType(int referencedComponentType) {
+		Builder referencedComponentType(short referencedComponentType) {
 			this.referencedComponentType = referencedComponentType;
 			return getSelf();
 		}
@@ -399,7 +399,7 @@ public class SnomedConceptDocument extends SnomedComponentDocument implements IT
 	private final boolean primitive;
 	private final boolean exhaustive;
 	private final SnomedRefSetType refSetType;
-	private final int referencedComponentType;
+	private final short referencedComponentType;
 	private final int mapTargetComponentType;
 	private final boolean structural;
 	private final long refSetStorageKey;
@@ -421,7 +421,7 @@ public class SnomedConceptDocument extends SnomedComponentDocument implements IT
 			final boolean primitive,
 			final boolean exhaustive, 
 			final SnomedRefSetType refSetType, 
-			final int referencedComponentType,
+			final short referencedComponentType,
 			final int mapTargetComponentType,
 			final long refSetStorageKey,
 			final boolean structural,
@@ -492,7 +492,7 @@ public class SnomedConceptDocument extends SnomedComponentDocument implements IT
 		return refSetType;
 	}
 	
-	public int getReferencedComponentType() {
+	public short getReferencedComponentType() {
 		return referencedComponentType;
 	}
